with windows shell prelaunch hook to found app hook and simplyfied it
